fix gluon files
