+++ tests/gluon/test_gluon_model_export.py
@@ -23,7 +23,7 @@ from mlflow.utils.environment import _mlflow_conda_env
 from mlflow.utils.file_utils import TempDir
 from mlflow.utils.model_utils import _get_flavor_configuration
 
-from tests.gluon_autolog.test_gluon_autolog import get_estimator
+from tests.gluon.utils import get_estimator
 from tests.helper_functions import (
     pyfunc_serve_and_score_model,
     _compare_conda_env_requirements,
